chore: refactor CLP Analyzer
